Extract FeedLoaderStub into a shared scope to remove duplication
